backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.59.2
pydantic==2.5.2
python-decouple==3.8
langchain==0.1.12
langchain-openai==0.2.14
beautifulsoup4==4.12.3
requests==2.31.0
chromadb==0.4.24 
tiktoken==0.8.0
pdfplumber==0.11.5
gunicorn==23.0.0
python-multipart==0.0.20